DIGEQ-159 convert to float to allow decimals

+++ validators/equal.py
@@ -8,7 +8,7 @@ class Equal(Numeric):
 
     def is_valid(self, response):
         value = self._schema['value']
-        return not value or super(Equal, self).is_valid(response)and int(response) != int(value)
+        return not value or super(Equal, self).is_valid(response) and float(response) != float(value)
 
     def get_message(self, response):
         value = self._schema['value']
